pyproject.toml
[project]
name = "winipedia-pyside"
version = "0.2.40"
description = "A package for pyside utils"
authors = [
    {name = "Winipedia",email = "win.steveker@gmx.de"}
]
license = "MIT"
readme = "README.md"
requires-python = ">=3.12,<3.14"
dynamic = ["dependencies"]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[[tool.poetry.packages]]
include = "winipedia_pyside"

[tool.poetry.group.dev.dependencies]
pytest-qt = "*"
types-pyyaml = "*"
pytest = "*"
ruff = "*"
types-defusedxml = "*"
mypy = "*"
pytest-mock = "*"
types-tqdm = "*"
pre-commit = "*"
types-setuptools = "*"
bandit = "*"
types-networkx = "*"


[tool.poetry.dependencies]
pyqt-toast-notification = "*"
pyside6 = "*"
winipedia-utils = "*"
[tool.ruff]
exclude = [".*", "**/migrations/*.py"]

[tool.ruff.lint]
select = ["ALL"]
ignore = ["D203", "D213", "COM812", "ANN401"]
fixable = ["ALL"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.mypy]
strict = true
warn_unreachable = true
show_error_codes = true
files = "."

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.bandit]
exclude_dirs = ["experiment.py"]
